services/auth/requirements.txt
psycopg2-binary==2.9.11
bcrypt==4.1.1
python-dotenv==1.0.0
fastapi==0.110.2
uvicorn[standard]==0.29.0